+++ src/haddock/modules/sampling/lightdock/__init__.py
@@ -2,12 +2,13 @@
 import logging
 import shutil
 import subprocess
+
 from pathlib import Path
-from haddock.modules import BaseHaddockModule
-from haddock.ontology import Format, ModuleIO, PDBFile
-from haddock.pdbutil import PDBFactory
-from haddock.modules import working_directory
+
 from haddock.defaults import NUM_CORES
+from haddock.libs.libpdb import PDBFactory
+from haddock.modules import BaseHaddockModule, working_directory
+from haddock.ontology import Format, ModuleIO, PDBFile
 
 
 logger = logging.getLogger(__name__)
